front_end_example/full_code/chatbot_backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 24.9.2
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.13=pypi_0
aiosignal=1.3.2=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.8.0=pypi_0
attrs=25.3.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.2.25=haa95532_0
cachecontrol=0.14.2=pypi_0
cachetools=5.5.2=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
colorama=0.4.6=pypi_0
cryptography=44.0.2=pypi_0
dataclasses-json=0.6.7=pypi_0
distro=1.9.0=pypi_0
dnspython=2.7.0=pypi_0
docstring-parser=0.16=pypi_0
email-validator=2.2.0=pypi_0
expat=2.6.4=h8ddb27b_0
fastapi=0.115.11=pypi_0
fastapi-cli=0.0.7=pypi_0
firebase-admin=6.7.0=pypi_0
frozenlist=1.5.0=pypi_0
google-api-core=2.24.2=pypi_0
google-api-python-client=2.164.0=pypi_0
google-auth=2.38.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-cloud-aiplatform=1.71.1=pypi_0
google-cloud-bigquery=3.30.0=pypi_0
google-cloud-core=2.4.3=pypi_0
google-cloud-firestore=2.20.1=pypi_0
google-cloud-resource-manager=1.14.1=pypi_0
google-cloud-storage=2.19.0=pypi_0
google-crc32c=1.6.0=pypi_0
google-resumable-media=2.7.2=pypi_0
googleapis-common-protos=1.69.1=pypi_0
greenlet=3.1.1=pypi_0
grpc-google-iam-v1=0.14.1=pypi_0
grpcio=1.71.0=pypi_0
grpcio-status=1.71.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httplib2=0.22.0=pypi_0
httptools=0.6.4=pypi_0
httpx=0.28.1=pypi_0
httpx-sse=0.4.0=pypi_0
idna=3.10=pypi_0
jinja2=3.1.6=pypi_0
jiter=0.9.0=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
langchain=0.3.20=pypi_0
langchain-community=0.3.19=pypi_0
langchain-core=0.3.45=pypi_0
langchain-google-vertexai=2.0.15=pypi_0
langchain-openai=0.3.8=pypi_0
langchain-text-splitters=0.3.6=pypi_0
langgraph=0.3.10=pypi_0
langgraph-checkpoint=2.0.19=pypi_0
langgraph-prebuilt=0.1.3=pypi_0
langgraph-sdk=0.1.57=pypi_0
langsmith=0.3.13=pypi_0
libffi=3.4.4=hd77b12b_1
libmpdec=4.0.0=h827c3e9_0
markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=pypi_0
marshmallow=3.26.1=pypi_0
mdurl=0.1.2=pypi_0
msgpack=1.1.0=pypi_0
multidict=6.1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
numpy=2.2.3=pypi_0
openai=1.66.3=pypi_0
openssl=3.0.16=h3f729d1_0
orjson=3.10.15=pypi_0
packaging=24.2=pypi_0
pip=25.0=py313haa95532_0
propcache=0.3.0=pypi_0
proto-plus=1.26.1=pypi_0
protobuf=5.29.3=pypi_0
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.1=pypi_0
pycparser=2.22=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pydantic-settings=2.8.1=pypi_0
pygments=2.19.1=pypi_0
pyjwt=2.10.1=pypi_0
pyparsing=3.2.1=pypi_0
pypdf=5.3.1=pypi_0
python=3.13.2=hadb2040_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-multipart=0.0.20=pypi_0
python_abi=3.13=0_cp313
pyyaml=6.0.2=pypi_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
requests-toolbelt=1.0.0=pypi_0
rich=13.9.4=pypi_0
rich-toolkit=0.13.2=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
setuptools=75.8.0=py313haa95532_0
shapely=2.0.7=pypi_0
shellingham=1.5.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
sqlalchemy=2.0.39=pypi_0
sqlite=3.45.3=h2bbff1b_0
starlette=0.46.1=pypi_0
tenacity=9.0.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tiktoken=0.9.0=pypi_0
tk=8.6.14=h0416ee5_0
tqdm=4.67.1=pypi_0
typer=0.15.2=pypi_0
typing-extensions=4.12.2=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2025a=h04d1e81_0
uritemplate=4.1.1=pypi_0
urllib3=2.3.0=pypi_0
uvicorn=0.34.0=pypi_0
vc=14.42=haa95532_4
vertexai=1.71.1=pypi_0
vs2015_runtime=14.42.34433=he0abc0d_4
watchfiles=1.0.4=pypi_0
websockets=15.0.1=pypi_0
wheel=0.45.1=py313haa95532_0
xz=5.6.4=h4754444_1
yarl=1.18.3=pypi_0
zlib=1.2.13=h8cc25b3_1
zstandard=0.23.0=pypi_0
